Minor clean-up/refactoring.
